rockchip: video: mipi: Add rk3288 soc specific driver for mipi dsi
Add rk3288 soc specific driver for mipi dsi.

Acked-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>

+++ drivers/video/rockchip/rk_mipi.h
@@ -8,7 +8,7 @@
 #define __RK_MIPI_H
 
 struct rk_mipi_priv {
-	ulong regs;
+	uintptr_t regs;
 	void *grf;
 	struct udevice *panel;
 	struct mipi_dsi *dsi;
